+++ brane-drv/src/lib.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    26 Sep 2022, 12:00:46
 //  Last edited:
-//    28 Nov 2022, 16:09:14
+//    12 Jul 2023, 16:31:18
 //  Auto updated?
 //    Yes
 // 
@@ -20,4 +20,5 @@ pub mod errors;
 pub mod spec;
 pub mod planner;
 pub mod vm;
+pub mod gc;
 pub mod handler;